+++ packages/tracing/src/browser/web-vitals/onFID.ts
