+++ packages/tracing/src/browser/web-vitals/onLCP.ts
